added validation for uniqueness of campsite name; added shoulda/model testing
