+++ app/models/site.rb
@@ -1,3 +1,4 @@
 class Site < ActiveRecord::Base
   belongs_to :vendor
+  validates :name, uniqueness: true
 end
