+++ app/views/sites/new.html.erb
@@ -1,3 +1,14 @@
+<% if @site.errors.any? %>
+  <section id='errors'>
+    <ul>
+      <% @site.errors.full_messages.each do |message|%>
+        <li><%= message %></li>
+      <% end %>
+    </ul>
+  </section>
+<% end %>
+
+
 <%= form_for Site.new do |f| %>
   <%= f.label :name %>
   <%= f.text_field :name %>
