+++ spec/models/vendor_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+describe Vendor, type: :model do
+  it { is_expected.to have_many :sites }
+end
